dbt_dagster/requirements.txt
dagster_dbt==0.16.17
dbt-bigquery==1.3.0
dagit==1.0.17
dagster-postgres==0.16.17
pandas==1.5.1